add readme and run configuration for rest sample
