+++ .run/general-rest-go-client.run.xml
@@ -1,0 +1,15 @@
+<component name="ProjectRunConfigurationManager">
+    <configuration default="false" name="general-rest-go-client" type="GoApplicationRunConfiguration" factoryName="Go Application" singleton="false">
+        <module name="dubbo-go-samples" />
+        <working_directory value="$PROJECT_DIR$" />
+        <envs>
+            <env name="APP_LOG_CONF_FILE" value="$PROJECT_DIR$/general/rest/go-client/conf/log.yml" />
+            <env name="CONF_CONSUMER_FILE_PATH" value="$PROJECT_DIR$/general/rest/go-client/conf/client.yml" />
+        </envs>
+        <kind value="PACKAGE" />
+        <filePath value="$PROJECT_DIR$/general/rest/go-client/cmd/client.go" />
+        <package value="github.com/apache/dubbo-go-samples/general/rest/go-client/cmd" />
+        <directory value="$PROJECT_DIR$" />
+        <method v="2" />
+    </configuration>
+</component>
