+++ .run/general-rest-go-server.run.xml
@@ -1,0 +1,15 @@
+<component name="ProjectRunConfigurationManager">
+    <configuration default="false" name="general-rest-go-server" type="GoApplicationRunConfiguration" factoryName="Go Application" singleton="false">
+        <module name="dubbo-go-samples" />
+        <working_directory value="$PROJECT_DIR$" />
+        <envs>
+            <env name="APP_LOG_CONF_FILE" value="$PROJECT_DIR$/general/rest/go-server/conf/log.yml" />
+            <env name="CONF_PROVIDER_FILE_PATH" value="$PROJECT_DIR$/general/rest/go-server/conf/server.yml" />
+        </envs>
+        <kind value="PACKAGE" />
+        <filePath value="$PROJECT_DIR$/general/rest/go-server/cmd/server.go" />
+        <package value="github.com/apache/dubbo-go-samples/general/rest/go-server/cmd" />
+        <directory value="$PROJECT_DIR$" />
+        <method v="2" />
+    </configuration>
+</component>
