+++ general/rest/go-client/cmd/version.go
