+++ general/rest/go-server/cmd/version.go
